Support better templating of the Cargo.toml
The one openapi-generator generates always has to be edited; so I came
up with this solution for now.

Usage: This will pull down the specs and build a fresh client based on
them and the latest openapi-generator:

VERSION=x.x.x make ['central' or 'service']

Notably, this template will need to be maintained to manage authors and
dependencies.

Signed-off-by: Erik Hollensbe <linux@hollensbe.org>

Makefile: update to support descriptions

Signed-off-by: Erik Hollensbe <linux@hollensbe.org>

meh

Signed-off-by: Erik Hollensbe <linux@hollensbe.org>

asdf

Signed-off-by: Erik Hollensbe <linux@hollensbe.org>